packages/bu-sfdc-integrations/pyproject.toml
[tool.poetry]
name = "bu-sfdc-integrations"
version = "0.1.0"
description = ""
authors = ["juslu-betterup <122490701+juslu-betterup@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "bu_sfdc_integrations"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
click-log = "^0.4.0"
simple-salesforce = "^1.12.3"
attrs = "^22.2.0"
numpy = "^1.24.1"
snowflake-sqlalchemy = "^1.4.4"

[tool.poetry.group.dev.dependencies]
requests = "^2.28.2"
urllib3 = "^1.26.14"
pytest = "^7.2.1"
pytest-mock = "^3.10.0"
mock = "^5.0.1"

[tool.poetry.scripts]
update-sfdc-contacts = "bu_sfdc_integrations.contacts_updater:main"
update-sfdc-accounts = "bu_sfdc_integrations.accounts_updater:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
